Move repair calculations to core
